+++ server/index.js
@@ -5,6 +5,7 @@ const user =require('./resolvers/user-resolver')
 const UserSchema=require('./models/user')
 const CompanySchema=require('./models/company')
 const company=require('./resolvers/company-resolver')
+const item=require('./resolvers/item-resolver')
 const auth=require('./token.js')
 const cookieParser = require('cookie-parser')
 const multer=require('multer')
@@ -83,15 +84,13 @@ app.post('/user/logout', auth.verify, user.logout)
 app.get('/user/get/:userId/:key', user.keyVerification, user.getUserById)
 
 
-app.get('/item/get/:itemId', user.getItemInfo)
-app.get('/item-transaction/get/:itemId', user.getItemTransactions)
 app.get('/pending-trade/get', user.getPendingTrades)
 app.post('/trade/create',auth.verify, user.createPendingTrade)
 app.post('/trade/complete', auth.verify, user.completeTrade)
 app.get('/completed-trade/get', auth.verify, user.getCompletedTrades)
 app.get('/completed-trade/get/:userId/:key', user.keyVerification, user.getCompletedTrades)
-app.post('/profile-pic/upload/:itemId', multer().single('file'), user.uploadProfilePic)
-app.get('/profile-pic/get/:itemId', user.getProfilePic)
+
+
 
 app.post('/company_register', company.register)
 app.post('/company_login', company.login)
@@ -100,11 +99,17 @@ app.get(`/inventory/:id`, auth.verify, company.getCompany)
 app.post('/company/createItem', company.createItem)
 app.post('/company/sellItem', company.sellItem)
 app.post('/company_addItem', auth.verify, company.addItem)
-app.post('/generate/qrcode/item', company.generateItemQRCode)
+
+app.get('/item/get/:itemId', item.getItemInfo)
+app.get('/item-transactions/get/:itemId', item.getItemTransactions)
+app.post('/generate/qrcode/item', item.generateItemQRCode)
+app.post('/profile-pic/upload/:itemId', multer().single('file'), item.uploadProfilePic)
+app.get('/profile-pic/get/:itemId', item.getProfilePic)
 app.get('/qrcode/item/:itemId', (req, res)=>{
     const itemId=req.params.itemId
     return res.sendFile(__dirname+`/images/item-qrcodes/${itemId}.png`)
 })
 app.get('/qrcode/profile/:userId', user.getProfileQRCode)
 app.post('/qrcode/scan', multer().single('file'), user.scanQrCode)
+
 app.listen(PORT, ()=>{console.log(`app is listening in ${PORT}`)})
